requirements.txt
pandas==1.4.2
Pillow==9.3.0
plotly_express==0.4.1
streamlit==1.16.0
